Refactor tests directory structure
Move all test app related code to fixtures folder
