+++ .gitignore
@@ -11,5 +11,5 @@ phpunit.xml
 .phpunit.result.cache
 
 # Tests
-tests/config/reference.php
-tests/.var/
+tests/Fixtures/config/reference.php
+tests/Fixtures/.var/
