+++ phpunit.xml.dist
@@ -9,7 +9,7 @@
          failOnWarning="true"
 >
     <php>
-        <server name="KERNEL_CLASS" value="League\Bundle\OAuth2ServerBundle\Tests\TestKernel" />
+        <server name="KERNEL_CLASS" value="League\Bundle\OAuth2ServerBundle\Tests\Fixtures\TestKernel" />
         <env name="SYMFONY_DEPRECATIONS_HELPER" value="max[self]=0&amp;ignoreFile=./tests/baseline-ignore" />
     </php>
