+++ tests/Acceptance/CustomPersistenceManagerTest.php
@@ -16,13 +16,13 @@ use League\Bundle\OAuth2ServerBundle\Model\Client;
 use League\Bundle\OAuth2ServerBundle\Model\RefreshToken;
 use League\Bundle\OAuth2ServerBundle\OAuth2Events;
 use League\Bundle\OAuth2ServerBundle\Service\CredentialsRevokerInterface;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeAccessTokenManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeAuthorizationCodeManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeClientManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeCredentialsRevoker;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeDeviceCodeManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeRefreshTokenManager;
 use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FixtureFactory;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeAccessTokenManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeAuthorizationCodeManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeClientManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeDeviceCodeManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeRefreshTokenManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Service\FakeCredentialsRevoker;
 use League\Bundle\OAuth2ServerBundle\Tests\TestHelper;
 use League\Bundle\OAuth2ServerBundle\ValueObject\RedirectUri;
 use PHPUnit\Framework\MockObject\MockObject;
