+++ tests/Fixtures/Grant/FakeGrantNullAccessTokenTTL.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures;
+namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Grant;
 
 use League\OAuth2\Server\Grant\AbstractGrant;
 use League\OAuth2\Server\Grant\GrantTypeInterface;
@@ -10,7 +10,7 @@ use League\OAuth2\Server\ResponseTypes\ResponseTypeInterface;
 use Nyholm\Psr7\Response;
 use Psr\Http\Message\ServerRequestInterface;
 
-final class FakeGrantUndefinedAccessTokenTTL extends AbstractGrant implements GrantTypeInterface
+final class FakeGrantNullAccessTokenTTL extends AbstractGrant implements GrantTypeInterface
 {
     public function getIdentifier(): string
     {
