+++ tests/Fixtures/Grant/FakeLegacyGrant.php
@@ -2,23 +2,32 @@
 
 declare(strict_types=1);
 
-namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures;
+namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Grant;
 
+use League\Bundle\OAuth2ServerBundle\AuthorizationServer\GrantTypeInterface;
 use League\OAuth2\Server\Grant\AbstractGrant;
-use League\OAuth2\Server\Grant\GrantTypeInterface;
 use League\OAuth2\Server\ResponseTypes\ResponseTypeInterface;
 use Nyholm\Psr7\Response;
 use Psr\Http\Message\ServerRequestInterface;
 
-final class FakeGrant extends AbstractGrant implements GrantTypeInterface
+// TODO remove code bloc when bundle interface and configurator will be deleted
+/**
+ * @deprecated
+ */
+final class FakeLegacyGrant extends AbstractGrant implements GrantTypeInterface
 {
     public function getIdentifier(): string
     {
-        return 'fake_grant';
+        return 'fake_legacy_grant';
     }
 
     public function respondToAccessTokenRequest(ServerRequestInterface $request, ResponseTypeInterface $responseType, \DateInterval $accessTokenTTL): ResponseTypeInterface
     {
         return new Response();
+    }
+
+    public function getAccessTokenTTL(): ?\DateInterval
+    {
+        return new \DateInterval('PT5H');
     }
 }
