+++ tests/Fixtures/Manager/FakeClientManager.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures;
+namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager;
 
 use League\Bundle\OAuth2ServerBundle\Manager\ClientFilter;
 use League\Bundle\OAuth2ServerBundle\Manager\ClientManagerInterface;
