+++ tests/Fixtures/TestKernel.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace League\Bundle\OAuth2ServerBundle\Tests;
+namespace League\Bundle\OAuth2ServerBundle\Tests\Fixtures;
 
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use League\Bundle\OAuth2ServerBundle\LeagueOAuth2ServerBundle;
