+++ tests/Fixtures/config/services_custom_persistence.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeAccessTokenManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeAuthorizationCodeManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeClientManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeCredentialsRevoker;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeDeviceCodeManager;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeRefreshTokenManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeAccessTokenManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeAuthorizationCodeManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeClientManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeDeviceCodeManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Manager\FakeRefreshTokenManager;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Service\FakeCredentialsRevoker;
 
 return static function (ContainerConfigurator $container): void {
     $container->services()
