+++ tests/Fixtures/config/services_fake_grant.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeGrant;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeGrantNullAccessTokenTTL;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FakeGrantUndefinedAccessTokenTTL;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Grant\FakeGrant;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Grant\FakeGrantNullAccessTokenTTL;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Grant\FakeGrantUndefinedAccessTokenTTL;
 
 return static function (ContainerConfigurator $container): void {
     $container->services()
