+++ tests/Unit/OAuth2TokenTest.php
@@ -6,7 +6,7 @@ namespace League\Bundle\OAuth2ServerBundle\Tests\Unit;
 
 use League\Bundle\OAuth2ServerBundle\Security\Authentication\Token\OAuth2Token;
 use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\FixtureFactory;
-use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\User;
+use League\Bundle\OAuth2ServerBundle\Tests\Fixtures\Security\User;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Security\Core\Authorization\Voter\CacheableVoterInterface;
